requirements.txt
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
langchain==0.2.3
langchain-experimental==0.0.62
langchain-text-splitters==0.2.1
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.33.0
pydantic==2.7.3
pydantic_core==2.18.4
requests==2.32.3
requests-oauthlib==2.0.0
